refactor: Additional API updates to complement UI changes
