+++ uvdat/core/rest/__init__.py
@@ -2,6 +2,7 @@ from .chart import ChartViewSet
 from .data import RasterDataViewSet, VectorDataViewSet
 from .dataset import DatasetViewSet
 from .layer import LayerFrameViewSet, LayerViewSet
+from .networks import NetworkViewSet
 from .project import ProjectViewSet
 from .regions import RegionViewSet
 from .simulations import SimulationViewSet
@@ -15,6 +16,7 @@ __all__ = [
     RasterDataViewSet,
     VectorDataViewSet,
     DatasetViewSet,
+    NetworkViewSet,
     RegionViewSet,
     SimulationViewSet,
     UserViewSet,
