+++ uvdat/urls.py
@@ -12,6 +12,7 @@ from uvdat.core.rest import (
     DatasetViewSet,
     LayerFrameViewSet,
     LayerViewSet,
+    NetworkViewSet,
     ProjectViewSet,
     RasterDataViewSet,
     RegionViewSet,
@@ -38,6 +39,8 @@ router.register(r'rasters', RasterDataViewSet, basename='rasters')
 router.register(r'vectors', VectorDataViewSet, basename='vectors')
 router.register(r'source-regions', RegionViewSet, basename='source-regions')
 router.register(r'simulations', SimulationViewSet, basename='simulations')
+router.register(r'networks', NetworkViewSet, basename='networks')
+
 
 urlpatterns = [
     path('accounts/', include('allauth.urls')),
